keep only Workflow component and move it into scaffolder-react
