+++ plugins/scaffolder-react/package.json
@@ -57,6 +57,7 @@
     "json-schema": "^0.4.0",
     "json-schema-library": "^7.3.9",
     "lodash": "^4.17.21",
+    "react-use": "^17.2.4",
     "qs": "^6.9.4",
     "zen-observable": "^0.10.0",
     "zod": "~3.18.0",
