+++ plugins/scaffolder-react/src/next/components/index.ts
@@ -17,3 +17,4 @@ export * from './Stepper';
 export * from './TemplateCard';
 export * from './ReviewState';
 export * from './TemplateGroup';
+export * from './Workflow';
